+++ driver/internal/protocol/auth/scrampbkdf2sha256.go
@@ -37,9 +37,10 @@ func (a *SCRAMPBKDF2SHA256) Typ() string { return MtSCRAMPBKDF2SHA256 }
 func (a *SCRAMPBKDF2SHA256) Order() byte { return MoSCRAMPBKDF2SHA256 }
 
 // PrepareInitReq implements the Method interface.
-func (a *SCRAMPBKDF2SHA256) PrepareInitReq(prms *Prms) {
+func (a *SCRAMPBKDF2SHA256) PrepareInitReq(prms *Prms) error {
 	prms.addString(a.Typ())
 	prms.addBytes(a.clientChallenge)
+	return nil
 }
 
 // InitRepDecode implements the Method interface.